internal: Add new 'MinusOne' family of traits and functions
These represents integer types that have -1 values. There are both
signed and unsigned variants, and it is possible to either limit usage
to signed or unsigned types, or to allow use of either.

Note that the -1 values for unsigned types are obtained by taking the -1
value for the corresponding signed type and casting it to the unsigned
type. (In most cases, this will correspond to T::MAX.)

This is useful in several cases where -1 has a special meaning in either
return values or arguments of functions.