resource: Make proc_rlimit()'s handling of new limits more consistent
Don't fall back on getrlimit() and setrlimit() for the current process
if it's not possible to set rlimits for arbitrary processes.

This allows "probing" to determine if it is possible to set rlimits for
other processes by calling something like this:

proc_rlimit(0, Resource::NOFILE,
    Some(getrlimit(Resource::NOFILE).unwrap()))

and checking for the ENOTSUP error.